feat: moved all client service into backend
