+++ src/components/PhoneInfo.tsx
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
-import { helpRequestService } from '@/lib/service';
 import { useSession } from '@/context/SessionProvider';
 import { HelpRequestData } from '../types/Requests';
+import { helpRequestService } from '../lib/actions';
 
 type PhoneInfoProps = {
   caseInfo: HelpRequestData;
