+++ src/types/Requests.ts
@@ -30,6 +30,22 @@ export type HelpRequestResources = {
   availability?: string[];
 };
 
+export type PuntoDeEntrega = {
+  name: string;
+  location: string;
+  city: string;
+  contact_name: string | null;
+  contact_phone: string | null | undefined;
+  contact_email: string | null;
+  vehicle_type: string | null;
+  cargo_type: string | null;
+  schedule: string | null;
+  additional_info: string | null;
+  latitude: number | null;
+  longitude: number | null;
+  status: string;
+};
+
 export type OmitSelect<T, K extends string> = { [key in keyof Omit<T, K>]: T[key] };
 export type SelectStringBuilder<T, K extends string> = { [key in keyof Omit<T, K>]: true };
